moving composite manipulators into item factory
